feat/ styled pages
