+++ src/components/Footer/Footer.tsx
@@ -1,8 +1,7 @@
-import { FC } from "react";
 import s from "./Footer.module.css";
 import { InstagramOutlined } from "@ant-design/icons";
 
-export const Footer: FC = () => {
+export const Footer = () => {
   return (
     <div className={s.footer}>
       <a className={s.name}>alexxandrrar </a>
