+++ src/pages/LoginationPage/LoginationPage.module.css
@@ -1,0 +1,4 @@
+.form {
+  align-items: center;
+  padding: 33vh 0 33vh 0;
+}
