+++ src/pages/SettingsPage/SettingsPage.module.css
@@ -1,0 +1,5 @@
+.title {
+  font-weight: 600;
+  font-size: 15px;
+  line-height: 32px;
+}
