+++ src/pages/SettingsPage/SettingsPage.tsx
@@ -1,7 +1,15 @@
+import { Skeleton, Divider } from "antd";
+import s from "./SettingsPage.module.css";
 export const SettingsPage = () => {
   return (
     <div>
-      <h1>Settings page</h1>
+      <h1 className={s.title}>SETTINGS</h1>
+      <Skeleton active />
+      <Skeleton active />
+      <Divider />
+      <Skeleton active />
+      <Divider />
+      <Skeleton active />
     </div>
   );
 };
